Add files via upload
This GitHub project is a simple portfolio card showcasing a basic yet stylish design created using HTML and CSS. The project focuses on creating a visually appealing representation of an individual's or a project's portfolio. The HTML file provides the structure, while the CSS file adds styling to enhance the overall look and feel. Users can easily understand and customize the code to personalize their own portfolio cards, making it a great starting point for individuals looking to create a clean and minimalist online presence.